Finish ecology proof-pack toolchain under tools/proofs/ecology
